+++ source/archetypes/assassin.js
@@ -5,6 +5,13 @@ const { getEmoji, getResistances } = require("../util/elementUtil");
 module.exports = new ArchetypeTemplate("Assassin",
 	"They'll be able to predict which combatants will critically hit and assess combatant elemental affinities, lining up massive damage with thier Daggers.",
 	"Wind",
+	{
+		maxHPGrowth: 25,
+		powerGrowth: 5,
+		speedGrowth: 0.5,
+		critRateGrowth: 1,
+		poiseGrowth: 0.25
+	},
 	["Daggers", "Cloak"],
 	(embed, adventure) => {
 		adventure.room.enemies.filter(combatant => combatant.hp > 0).concat(adventure.delvers).forEach(combatant => {