+++ source/archetypes/detective.js
@@ -5,6 +5,13 @@ const { getEmoji, getResistances } = require("../util/elementUtil");
 module.exports = new ArchetypeTemplate("Detective",
 	"They'll be able to predict which combatants will critically hit and assess combatant elemental affinities and even induce weaknesses on foes with their Sabotage Kit.",
 	"Earth",
+	{
+		maxHPGrowth: 25,
+		powerGrowth: 5,
+		speedGrowth: 0.5,
+		critRateGrowth: 1,
+		poiseGrowth: 0.25
+	},
 	["Pistol", "Sabotage Kit"],
 	(embed, adventure) => {
 		adventure.room.enemies.filter(combatant => combatant.hp > 0).concat(adventure.delvers).forEach(combatant => {