+++ source/archetypes/legionnaire.js
@@ -4,6 +4,13 @@ const { listifyEN } = require("../util/textUtil");
 module.exports = new ArchetypeTemplate("Legionnaire",
 	"They'll be able to predict who enemies are targeting with which moves. They'll be able to coodinate for big damage by inflicting Exposed on foes with their Shortsword.",
 	"Fire",
+	{
+		maxHPGrowth: 25,
+		powerGrowth: 5,
+		speedGrowth: 0.5,
+		critRateGrowth: 1,
+		poiseGrowth: 0.25
+	},
 	["Shortsword", "Scutum"],
 	(embed, adventure) => {
 		adventure.room.moves.forEach(({ userReference, targets, name, priority }) => {