+++ source/rooms/battle-slimes.js
@@ -4,6 +4,7 @@ module.exports = new RoomTemplate("Slime Fight",
 	"@{adventure}",
 	"Some slimes and oozes approach...",
 	[
+		new ResourceTemplate("1", "internal", "levelsGained"),
 		new ResourceTemplate("25*n", "loot", "gold")
 	]
 ).addEnemy("@{adventure} Slime", "0.5*n")