+++ source/rooms/finalbattle-elkemist.js
@@ -1,7 +1,9 @@
-const { RoomTemplate } = require("../classes");
+const { RoomTemplate, ResourceTemplate } = require("../classes");
 
 module.exports = new RoomTemplate("A Northern Laboratory",
 	"Water",
 	"Flasks, beakers, vials and a myriad unknown substances line the innumerable tables and shelves of this room. An elk wanders from table to whiteboard to shelf, muttering various alchemical formuae to itself and taking absolutely no notice of the party.",
-	[]
+	[
+		new ResourceTemplate("5", "internal", "levelsGained")
+	]
 ).addEnemy("Elkemist", "1");